// book-network-ui/src/app/services/models/book-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookResponse {
  archived?: boolean;
  author?: string;
  bookCover?: string;
  id?: number;
  isbn?: string;
  ownerId?: number;
  ownerName?: string;
  rate?: number;
  shareable?: boolean;
  synopsis?: string;
  title?: string;
}